% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_from_gs.R
\name{dct_from_gs}
\alias{dct_from_gs}
\title{Import a DCT specification from a Google Sheets spreadsheet}
\usage{
dct_from_gs(
  gs,
  sheets = TRUE,
  path = NULL,
  localBackup = NULL,
  preventOverwriting = psyverse::opts$get("preventOverwriting"),
  encoding = psyverse::opts$get("encoding")
)
}
\arguments{
\item{gs}{The URL to the Google Sheets spreadsheet, its key, or the
object resulting from a call to \code{\link[googlesheets4:gs4_get]{googlesheets4::gs4_get()}}.}

\item{sheets}{\code{TRUE} to read all worksheets, or a vector with the
indices or names of the worksheets to read.}

\item{path}{The path to save the DCT specifications.}

\item{localBackup}{If pointing to a filename in an existing path,
the spreadsheet will be saved as \code{xlsx} file using
the \link[openxlsx:openxlsx]{openxlsx::openxlsx} package.}

\item{preventOverwriting}{Whether to prevent overwriting.}

\item{encoding}{The encoding to use.}
}
\value{
Invisibly, an object with the worksheets and the DCT objects.
}
\description{
Import a DCT specification from a Google Sheets spreadsheet
}
