% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_from_xlsx.R
\name{dct_from_xlsx}
\alias{dct_from_xlsx}
\title{Title}
\usage{
dct_from_xlsx(
  xlsx,
  path = NULL,
  preventOverwriting = psyverse::opts$get("preventOverwriting"),
  encoding = psyverse::opts$get("encoding")
)
}
\arguments{
\item{encoding}{}
}
\value{

}
\description{
Title
}
