% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dct_dir.R
\name{extract_dct_dir}
\alias{extract_dct_dir}
\title{Extract DCT specifications from all files in a directory}
\usage{
extract_dct_dir(path, extension = "dct", regex, delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE, silent = FALSE)
}
\arguments{
\item{path}{The path from where to read the DCT files.}

\item{extension}{The extension of the file; convenient alternative to
specifying a \code{regex}.}

\item{regex}{A regular expression; when provided, overrides the \code{extension}
argument to guide file selection in the \code{path}.}

\item{delimiterRegEx}{The regular expression specifying how the YAML fragments
specifying the constructs are delimited. Should normally never be changed.}

\item{ignoreOddDelimiters}{Whether to ignore a final odd delimiter, if
encountered.}

\item{silent}{Whether to provide information on progress.}
}
\value{
An object of class \code{dctRawSpecListSet} for processing by \code{\link[=parse_dct_specs]{parse_dct_specs()}}.
}
\description{
This function processes all DCT specifications in a directory.
}
\details{
This function is called by \code{\link[=process_dir]{process_dir()}}; it is normally not
necessary to call this function directly.
}
\examples{
\dontrun{extract_dct_dir("A:/path/to/some/directory");
}
}
