% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_id.R
\name{invert_id}
\alias{invert_id}
\title{Invert identifier}
\usage{
invert_id(x)
}
\arguments{
\item{x}{The identifier(s) as a character vector.}
}
\value{
The identifier(s) as a numeric vector.
}
\description{
Invert the identifier for one or more constructs.
}
\examples{
invert_id(generate_id('example'));
}
