% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_construct_overview.R,
%   R/generate_instruction_overview.R
\name{generate_construct_overview}
\alias{generate_construct_overview}
\alias{generate_instruction_overview}
\title{Generate construct overviews and instruction overviews}
\usage{
generate_construct_overview(
  dctSpec,
  headingLevel = 3,
  hyperlink_ucids = "Markdown",
  urlPrefix = "#"
)

generate_instruction_overview(
  dctSpecDf,
  type,
  headingLevel = 3,
  hyperlink_ucids = "Markdown",
  urlPrefix = "#"
)
}
\arguments{
\item{dctSpec}{The DCT specification, as resulting from a call
to \code{\link[=load_dct_specs]{load_dct_specs()}} or \code{\link[=load_dct_dir]{load_dct_dir()}}.}

\item{headingLevel}{The level of the heading in the Markdown output
that is produces.}

\item{hyperlink_ucids}{The type of hyperlinks to generate; must be
a valid string. Currently, if the value is"\code{Markdown}" or "\code{HTML}",
hyperlinks in the corresponding formats are produced, and if it is
"\code{none}" (or, actually any other string value), nothing is produced.}

\item{urlPrefix}{The prefix to insert before the URL in the produced
hyperlink. The default, "\verb{#}", results in a link to an
anchor (an HTML \code{a} element) on the current page.}

\item{dctSpecDf}{The DCT specification dataframer, as produced
by a call to \code{\link[=load_dct_specs]{load_dct_specs()}} or \code{\link[=load_dct_dir]{load_dct_dir()}}, and stored within
the resulting object.}

\item{type}{For instruction overviews, the type of instruction
to generate can be specified: must be}
}
\value{
A character string with the overview.
}
\description{
These functions use a DCT specification to generate a
construct overview or an instruction overview.
}
\examples{
### Add example
}
