% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dct_specs.R
\name{parse_dct_specs}
\alias{parse_dct_specs}
\title{Parse DCT specifications}
\usage{
parse_dct_specs(dctSpecs, headingLevel = 2)
}
\arguments{
\item{dctSpecs}{The DCT specifications (a list).}

\item{headingLevel}{The heading level for Rmd output.}
}
\value{
The object of parsed DCT specifications.
}
\description{
This function parses DCT specifications; it's normally
called by \code{\link[=load_dct_dir]{load_dct_dir()}} or \code{\link[=load_dct_specs]{load_dct_specs()}}, so you
won't have to use it directly.
}
