% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_yaml.R
\name{save_to_yaml}
\alias{save_to_yaml}
\title{Save a psyverse object or YAML character vector to a file}
\usage{
save_to_yaml(
  x,
  file,
  preventOverwriting = psyverse::opts$get("preventOverwriting"),
  encoding = psyverse::opts$get("encoding")
)
}
\arguments{
\item{x}{The object to save.}

\item{file}{The file to save to.}

\item{preventOverwriting}{Whether to prevent overwriting.}

\item{encoding}{The encoding to use. Note that in general, encoding seems
to have been invented primarily as a source of frustration, and it rarely
disappoints. If unsure, use UTF-8. If using UTF-8, the approach from
\url{https://kevinushey.github.io/blog/2018/02/21/string-encoding-and-r/} will
be used.}
}
\value{
The character vector that was written to the file.
}
\description{
Pretty much what it says on the box. But check the bit about encoding.
}
