% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soq_adapter_limonaid.R
\name{soq_adapter_limonaid}
\alias{soq_adapter_limonaid}
\title{SOQ adapter for LimeSurvey (using the {limonaid} package)}
\usage{
soq_adapter_limonaid(soq)
}
\arguments{
\item{soq}{The SOQ specification}
}
\value{
The .LSG content as {xml2} XML object.
}
\description{
SOQ adapter for LimeSurvey (using the {limonaid} package)
}
\examples{
\donttest{
UQID <- "eq60eng_7rs8g3bd";

all_soq_specs <-
  psyverse::read_SOQs_from_url();

selected_soq <- all_soq_specs[[UQID]];

lsgContent <-
  psyverse::soq_adapter_limonaid(
    selected_soq
  );
}

}
