% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dct_spec.R, R/load_dct_dir.R,
%   R/load_dct_specs.R
\docType{data}
\name{example_dct_spec}
\alias{example_dct_spec}
\alias{load_dct_dir}
\alias{load_dct_specs}
\alias{print.dct_specs}
\alias{plot.dct_specs}
\title{Load DCT specifications from a file or multiple files}
\format{An object of class \code{character} of length 1.}
\usage{
example_dct_spec

load_dct_dir(path, recursive = TRUE, extension = "rock|dct", regex,
  dctContainer = "dct", arrowDirection = "forward",
  delimiterRegEx = "^---$", ignoreOddDelimiters = FALSE,
  encoding = "UTF-8", silent = TRUE)

load_dct_specs(file, text, delimiterRegEx = "^---$",
  dctContainer = "dct", arrowDirection = "forward",
  ignoreOddDelimiters = FALSE, encoding = "UTF-8", silent = TRUE)

\method{print}{dct_specs}(x, ...)

\method{plot}{dct_specs}(x, ...)
}
\arguments{
\item{path}{The path containing the files to read.}

\item{recursive}{Whether to also process subdirectories (\code{TRUE})
or not (\code{FALSE}).}

\item{extension}{The extension of the files to read; files with other extensions will
be ignored. Multiple extensions can be separated by a pipe (\code{|}).}

\item{regex}{Instead of specifing an extension, it's also possible to specify a regular
expression; only files matching this regular expression are read. If specified, \code{regex}
takes precedece over \code{extension},}

\item{dctContainer}{The container of the DCT specifications in the YAML
fragments. Because only DCT specifications are read that are stored in
this container, the files can contain YAML fragments with other data, too,
without interfering with the parsing of the DCT specifications.}

\item{arrowDirection}{The direction of the arrows in the visual representation
of the distributed construct taxonomy; either \code{forward}, \code{back}, \code{both}, or \code{none}.}

\item{delimiterRegEx}{The regular expression used to locate YAML
fragments}

\item{ignoreOddDelimiters}{Whether to throw an error (FALSE) or
delete the last delimiter (TRUE) if an odd number of delimiters is
encountered.}

\item{encoding}{The encoding to use when calling \code{\link[=readLines]{readLines()}}. Set to
NULL to let \code{\link[=readLines]{readLines()}} guess.}

\item{silent}{Whether to be silent (TRUE) or informative (FALSE).}

\item{file, text}{Either specify a \code{file} to read with encoding \code{encoding}, which will
then be read using \code{\link[base:readLines]{base::readLines()}} (if specified, takes precedence over \code{text});
or specify a \code{text} to process, which should be a character vectors where every
element is a line of the original source (like provided \code{\link[base:readLines]{base::readLines()}});
although if a character vector of one element \emph{but} including at least one
newline character (\code{\\n}) is provided as \code{text}, this is split at the newline
characters using \code{\link[base:strsplit]{base::strsplit()}}.}

\item{x}{The parsed \code{parsed_dct} object.}

\item{...}{Any other arguments are passed to the print command.}
}
\value{
An object with the \link[DiagrammeR:DiagrammeR]{DiagrammeR::DiagrammeR} graph stored
in \code{output$graph} and the instructions in \code{output$instr}.
}
\description{
These function load DCT specifications from the YAML fragments
in one (\code{load_dct_specs}) or multiple files (\code{load_dct_dir}).
}
\details{
\code{load_dct_dir} simply identifies all files and then calls
\code{load_dct_specs} for each of them. \code{load_dct_specs} loads the
YAML fragments containing the DCT specifications using
\code{\link[yum:load_yaml_fragments]{yum::load_yaml_fragments()}} and then parses the DCT
specifications into a visual representation as a
\link[DiagrammeR:DiagrammeR]{DiagrammeR::DiagrammeR} graph and Markdown documents with
the instructions for creating measurement instruments or
manipulations, and for coding measurement instruments,
manipulations, or aspects of a construct.
}
\examples{
load_dct_specs(text=example_dct_spec);

\dontrun{
dct::load_dct_dir(path="A:/some/path");
}

}
\keyword{datasets}
