% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dct_specs.R
\name{parse_dct_specs}
\alias{parse_dct_specs}
\alias{print.parsed_dct}
\alias{plot.parsed_dct}
\title{Parse DCT specifications}
\usage{
parse_dct_specs(dctSpecs, arrowDirection = "forward")

\method{print}{parsed_dct}(x, ...)

\method{plot}{parsed_dct}(x, ...)
}
\arguments{
\item{dctSpecs}{A list of lists of DCT specifications (class
\code{dctRawSpecListSet}), a list of DCT specifications (class
\code{dtcRawSpecList}) or a DCT specification (class \code{dtcRawSpec}).}

\item{arrowDirection}{The direction of the arrows in the visual representation
of the distributed construct taxonomy; either \code{forward}, \code{back}, \code{both}, or \code{none}.}

\item{x}{The parsed \code{parsed_dct} object.}

\item{...}{Any other arguments are passed to the print command.}
}
\value{
An object with the \link[DiagrammeR:DiagrammeR]{DiagrammeR::DiagrammeR} graph stored
in \code{output$graph} and the instructions in \code{output$instr}.
}
\description{
This function parses DCT specifications into a visual
representation as a \link[DiagrammeR:DiagrammeR]{DiagrammeR::DiagrammeR} graph and
Markdown documents with the instructions for creating
measurement instruments or manipulations, and for coding
measurement instruments, manipulations, or aspects of
a construct.
}
\details{
This function is called by \code{\link[=process_dir]{process_dir()}}; it is normally not
necessary to call this function directly.
}
\examples{
extractedSpecs <- extract_dct_specs(text=unlist(strsplit(dct::example_dct_spec, '\\n')));
parse_dct_specs(extractedSpecs);
}
