% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dct_specs.R
\name{parse_dct_specs}
\alias{parse_dct_specs}
\title{Parse DCT specifications}
\usage{
parse_dct_specs(
  dctSpecs,
  headingLevel = 2,
  hyperlink_ucids = "Markdown",
  urlPrefix = "#"
)
}
\arguments{
\item{dctSpecs}{The DCT specifications (a list).}

\item{headingLevel}{The heading level for Markdown output.}

\item{hyperlink_ucids, urlPrefix}{Passed on to the
\code{\link[=generate_instruction_overview]{generate_instruction_overview()}} and \code{\link[=generate_construct_overview]{generate_construct_overview()}}
functions.}
}
\value{
The object of parsed DCT specifications.
}
\description{
This function parses DCT specifications; it's normally
called by \code{\link[=load_dct_dir]{load_dct_dir()}} or \code{\link[=load_dct_specs]{load_dct_specs()}}, so you
won't have to use it directly.
}
