% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dir.R
\name{process_dir}
\alias{process_dir}
\title{Process a directory with DCT files}
\usage{
process_dir(path, extension = "dct", regex, delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE, silent = FALSE)
}
\arguments{
\item{path}{The path from where to read the DCT files.}

\item{extension}{The extension of the file; convenient alternative to
specifying a \code{regex}.}

\item{regex}{A regular expression; when provided, overrides the \code{extension}
argument to guide file selection in the \code{path}.}

\item{delimiterRegEx}{The regular expression specifying how the YAML fragments
specifying the constructs are delimited. Should normally never be changed.}

\item{ignoreOddDelimiters}{Whether to ignore a final odd delimiter, if
encountered.}

\item{silent}{Whether to provide information on progress.}
}
\value{
The result of a call to \code{parse_dct_specs} on all matching files.
}
\description{
This function locates and parses all DCT files in a directory. These
files are then parsed into a DiagrammeR graph as well as lists of
instructions for development and coding.
}
